+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-## 4.1.2
+## 4.2.0
 ### General
-* Drag & Drop Capability added on example project.
+* Drag & Drop Capability added on example project. 
+* Added `generatePDFFromDocuments` method to create a PDF from a mix of PDFs and images in any order.
+* Improved UI for example project.
 
 ## 4.1.1
 ### Android